Moved the LettersPanel from layouts to components
